chore: fix eslint
