+++ .eslintrc.js
@@ -34,6 +34,9 @@ module.exports = {
     'import/first': 'error',
     'import/newline-after-import': 'error',
     'import/no-duplicates': 'error',
+    'import/no-unresolved': 'off',
+    'import/no-extraneous-dependencies': 'off',
+
   },
   globals: {
     React: 'writable',
